+++ test/controller.test.ts
@@ -810,25 +810,64 @@ describe('Controller', () => {
         expect(databaseSaveSpy).toHaveBeenCalledTimes(1);
     });
 
-    it('Controller stop, should reset runtime lookups', async () => {
+    it('Syncs runtime lookups', async () => {
         await controller.start();
+        // @ts-expect-error private
+        Device.devices.clear();
+        // @ts-expect-error private
+        Device.deletedDevices.clear();
+        // @ts-expect-error private
+        Group.groups.clear();
 
         await mockAdapterEvents['deviceJoined']({networkAddress: 129, ieeeAddr: '0x129'});
+        // @ts-expect-error private
+        expect(Device.devices.size).toStrictEqual(1);
+        // @ts-expect-error private
+        expect(Device.deletedDevices.size).toStrictEqual(0);
+        expect(Device.byIeeeAddr('0x129', false)).toBeInstanceOf(Device);
+        expect(Device.byIeeeAddr('0x128', false)).toBeUndefined();
+
         await mockAdapterEvents['deviceJoined']({networkAddress: 128, ieeeAddr: '0x128'});
         await mockAdapterEvents['deviceLeave']({networkAddress: 128, ieeeAddr: '0x128'});
-        controller.createGroup(1);
-        expect(Device.byIeeeAddr('0x129', false)).toBeInstanceOf(Device);
+        // @ts-expect-error private
+        expect(Device.devices.size).toStrictEqual(1);
+        // @ts-expect-error private
+        expect(Device.deletedDevices.size).toStrictEqual(1);
+        expect(Device.byIeeeAddr('0x128', false)).toBeUndefined();
         expect(Device.byIeeeAddr('0x128', true)).toBeInstanceOf(Device);
+
+        await mockAdapterEvents['deviceJoined']({networkAddress: 128, ieeeAddr: '0x128'});
+        // @ts-expect-error private
+        expect(Device.devices.size).toStrictEqual(2);
+        // @ts-expect-error private
+        expect(Device.deletedDevices.size).toStrictEqual(0);
+        const device2 = Device.byIeeeAddr('0x128', false);
+        expect(device2).toBeInstanceOf(Device);
+        expect(() => {
+            device2!.undelete();
+        }).toThrow(`Device '0x128' is not deleted`);
+
+        controller.createGroup(1);
+        // @ts-expect-error private
+        expect(Group.groups.size).toStrictEqual(1);
         expect(Group.byGroupID(1)).toBeInstanceOf(Group);
+        expect(Group.byGroupID(2)).toBeUndefined();
+
+        const group2 = controller.createGroup(2);
+        group2.removeFromNetwork();
+        // @ts-expect-error private
+        expect(Group.groups.size).toStrictEqual(1);
+        expect(Group.byGroupID(1)).toBeInstanceOf(Group);
+        expect(Group.byGroupID(2)).toBeUndefined();
 
         await controller.stop();
 
         // @ts-expect-error private
-        expect(Device.devices).toStrictEqual({});
+        expect(Device.devices.size).toStrictEqual(0);
         // @ts-expect-error private
-        expect(Device.deletedDevices).toStrictEqual({});
+        expect(Device.deletedDevices.size).toStrictEqual(0);
         // @ts-expect-error private
-        expect(Group.groups).toStrictEqual({});
+        expect(Group.groups.size).toStrictEqual(0);
     });
 
     it('Controller start', async () => {
@@ -2083,18 +2122,24 @@ describe('Controller', () => {
     it('Device announce event should update network address when different', async () => {
         await controller.start();
         await mockAdapterEvents['deviceJoined']({networkAddress: 129, ieeeAddr: '0x129'});
+        expect(controller.getDeviceByNetworkAddress(129)?.ieeeAddr).toStrictEqual('0x129');
         expect(events.deviceAnnounce.length).toBe(0);
         await mockAdapterEvents['deviceAnnounce']({networkAddress: 9999, ieeeAddr: '0x129'});
-        expect(controller.getDeviceByIeeeAddr('0x129').networkAddress).toBe(9999);
-        expect(controller.getDeviceByIeeeAddr('0x129').getEndpoint(1).deviceNetworkAddress).toBe(9999);
+        expect(controller.getDeviceByIeeeAddr('0x129')?.networkAddress).toBe(9999);
+        expect(controller.getDeviceByIeeeAddr('0x129')?.getEndpoint(1)?.deviceNetworkAddress).toBe(9999);
+        expect(controller.getDeviceByNetworkAddress(129)).toBeUndefined();
+        expect(controller.getDeviceByNetworkAddress(9999)?.ieeeAddr).toStrictEqual('0x129');
     });
 
     it('Network address event should update network address when different', async () => {
         await controller.start();
         await mockAdapterEvents['deviceJoined']({networkAddress: 129, ieeeAddr: '0x129'});
+        expect(controller.getDeviceByNetworkAddress(129)?.ieeeAddr).toStrictEqual('0x129');
         await mockAdapterEvents['networkAddress']({networkAddress: 9999, ieeeAddr: '0x129'});
         expect(controller.getDeviceByIeeeAddr('0x129').networkAddress).toBe(9999);
         expect(controller.getDeviceByIeeeAddr('0x129').getEndpoint(1).deviceNetworkAddress).toBe(9999);
+        expect(controller.getDeviceByNetworkAddress(129)).toBeUndefined();
+        expect(controller.getDeviceByNetworkAddress(9999)?.ieeeAddr).toStrictEqual('0x129');
     });
 
     it('Network address event shouldnt update network address when the same', async () => {
@@ -2408,13 +2453,16 @@ describe('Controller', () => {
         await controller.start();
         await mockAdapterEvents['deviceJoined']({networkAddress: 129, ieeeAddr: '0x129'});
         await mockAdapterEvents['deviceJoined']({networkAddress: 129, ieeeAddr: '0x129'});
+        expect(controller.getDeviceByNetworkAddress(129)?.ieeeAddr).toStrictEqual('0x129');
         expect(events.deviceJoined.length).toBe(1);
         expect(equalsPartial(events.deviceJoined[0].device, {ID: 2, networkAddress: 129, ieeeAddr: '0x129'})).toBeTruthy();
-        expect(controller.getDeviceByIeeeAddr('0x129').networkAddress).toBe(129);
+        expect(controller.getDeviceByIeeeAddr('0x129')?.networkAddress).toBe(129);
 
         await mockAdapterEvents['deviceJoined']({networkAddress: 130, ieeeAddr: '0x129'});
         expect(events.deviceJoined.length).toBe(1);
-        expect(controller.getDeviceByIeeeAddr('0x129').networkAddress).toBe(130);
+        expect(controller.getDeviceByIeeeAddr('0x129')?.networkAddress).toBe(130);
+        expect(controller.getDeviceByNetworkAddress(129)).toBeUndefined();
+        expect(controller.getDeviceByNetworkAddress(130)?.ieeeAddr).toStrictEqual('0x129');
     });
 
     it('Device joins and interview succeeds', async () => {
@@ -6995,7 +7043,7 @@ describe('Controller', () => {
             _skipDefaultResponse: false,
             meta: {configured: 1},
         });
-        expect((await controller.getGroups({})).length).toBe(2);
+        expect((await controller.getGroups()).length).toBe(2);
 
         const group1 = controller.getGroupByID(1);
         group1._members = Array.from(group1._members);